feat: 커스텀 에러를 상수 폴더로 이동
