+++ packages/shared/src/utils/error.ts
@@ -16,16 +16,3 @@ export function toErrorWithMessage(maybeError: unknown): Error {
 export function getErrorMeesage(error: unknown) {
   return toErrorWithMessage(error).message;
 }
-
-export class NoMemoError extends Error {
-  constructor() {
-    super('동일한 메모가 존재하지 않습니다.');
-    this.name = 'NoMemoError';
-  }
-}
-export class NoMemoListError extends Error {
-  constructor() {
-    super('메모 리스트가 존재하지 않습니다.');
-    this.name = 'NoMemoError';
-  }
-}
